효정 ChCategory.java ChCategoryRepository.java memberController.java Price.java PriceRepository.java -> Long을 integer로 수정 singUp.jsp 중

+++ src/main/java/hot/member/domain/Price.java
@@ -23,7 +23,7 @@ public class Price {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "price_no")
-	private Long priceNo;
+	private Integer priceNo;
 	
 	@Column(name = "price_amount")
 	private int priceAmount;
