+++ src/main/java/hot/member/repository/ChCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import hot.member.domain.ChCategory;
 
-public interface ChCategoryRepository extends JpaRepository<ChCategory, Long> {
+public interface ChCategoryRepository extends JpaRepository<ChCategory, Integer> {
 
 }
